requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cors==1.0.1
cryptography==45.0.3
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
future==1.0.0
gevent==25.5.1
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
is-disposable-email==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
setuptools==80.9.0
SQLAlchemy==2.0.41
tldextract==5.3.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
wtforms-validators==1.0.0
zope.event==5.0
zope.interface==7.2
scikit-learn==1.3.2
pandas==2.2.2

sentence-transformers
faiss-cpu
openai
